Bayes Factor for correlation test

Bayes Factor for correlation test

bf_corr_test(data, x, y, bf.prior = 0.707, ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The column in data containing the explanatory variable to be plotted on the x-axis. Can be entered either as a character string (e.g., "x") or as a bare expression (e.g, x).

y
The column in data containing the response (outcome) variable to be plotted on the y-axis. Can be entered either as a character string (e.g., "y") or as a bare expression (e.g, y).

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

...
Arguments passed on to bf_extractor

conf.level

Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

centrality

The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

conf.method

The type of index used for Credible Interval. Can be "hdi" (default), "eti", or "si" (see si(), hdi(), eti() functions from bayestestR package).

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

top.text

Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

output

If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

See also

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(tidyBF)

# to get dataframe
bf_corr_test(
  data = anscombe,
  x = x1,
  y = y4,
  bf.prior = 0.3,
  output = "dataframe"
)

#> # A tibble: 1 x 12
#>   term  estimate conf.low conf.high    pd rope.percentage prior.distribution
#>   <chr>    <dbl>    <dbl>     <dbl> <dbl>           <dbl> <chr>             
#> 1 rho     -0.182   -0.614     0.296 0.773           0.266 cauchy            
#>   prior.location prior.scale  bf10 method                        log_e_bf10
#>            <dbl>       <dbl> <dbl> <chr>                              <dbl>
#> 1              0         0.3 0.828 Bayesian correlation analysis     -0.189

# to get expression
bf_corr_test(
  data = anscombe,
  x = x1,
  y = y4,
  bf.prior = 0.8,
  output = "expression"
)

#> paste("log"["e"] * "(BF"["01"] * ") = " * "0.50" * ", ", widehat(italic(rho))["median"]^"posterior" * 
#>     " = " * "-0.23" * ", ", "CI"["95%"]^"HDI" * " [" * "-0.72" * 
#>     ", " * "0.30" * "], ", italic("r")["Cauchy"]^"JZS" * " = " * 
#>     "0.80")

# }