Bayes factor for random-effects meta-analysis

Bayes factor for random-effects meta-analysis

bf_meta_random(
  data,
  d = prior("norm", c(mean = 0, sd = 0.3)),
  tau = prior("invgamma", c(shape = 1, scale = 0.15)),
  k = 2L,
  conf.level = 0.95,
  output = "dataframe",
  top.text = NULL,
  ...
)

bf_meta(
  data,
  d = prior("norm", c(mean = 0, sd = 0.3)),
  tau = prior("invgamma", c(shape = 1, scale = 0.15)),
  k = 2L,
  conf.level = 0.95,
  output = "dataframe",
  top.text = NULL,
  ...
)

Arguments

data
A dataframe. It must contain columns named estimate (effect sizes or outcomes) and std.error (corresponding standard errors). These two columns will be used for y and SE arguments in metaBMA::meta_random.

d
prior distribution on the average effect size d. The prior probability density function is defined via prior.

tau
prior distribution on the between-study heterogeneity tau (i.e., the standard deviation of the study effect sizes dstudy in a random-effects meta-analysis. A (nonnegative) prior probability density function is defined via prior.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

output
If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

top.text
Text to display as top.text (will be displayed on top of the Bayes Factor top.text/message).

...
Arguments passed on to metaBMA::meta_random

labels

optional: character values with study labels. Can be a character vector or the quoted or unquoted name of the variable in data

rscale_contin

scale parameter of the JZS prior for the continuous covariates.

rscale_discrete

scale parameter of the JZS prior for discrete moderators.

centering

whether continuous moderators are centered.

logml

how to estimate the log-marginal likelihood: either by numerical integration ("integrate") or by bridge sampling using MCMC/Stan samples ("stan"). To obtain high precision with logml="stan", many MCMC samples are required (e.g., logml_iter=10000, warmup=1000).

summarize

how to estimate parameter summaries (mean, median, SD, etc.): Either by numerical integration (summarize = "integrate") or based on MCMC/Stan samples (summarize = "stan").

rel.tol

relative tolerance used for numerical integration using integrate. Use rel.tol=.Machine$double.eps for maximal precision (however, this might be slow).

logml_iter

number of iterations (per chain) from the posterior distribution of d and tau. The samples are used for computing the marginal likelihood of the random-effects model with bridge sampling (if logml="stan") and for obtaining parameter estimates (if summarize="stan"). Note that the argument iter=2000 controls the number of iterations for estimation of the random-effect parameters per study in random-effects meta-analysis.

silent_stan

whether to suppress the Stan progress bar.

Examples

# \donttest{
# setup
set.seed(123)
library(metaBMA)

# creating a dataframe
(df <-
  structure(
    .Data = list(
      study = c("1", "2", "3", "4", "5"),
      estimate = c(
        0.382047603321706,
        0.780783111514665,
        0.425607573765058,
        0.558365541235078,
        0.956473848429961
      ),
      std.error = c(
        0.0465576338644502,
        0.0330218199731529,
        0.0362834986178494,
        0.0480571500648261,
        0.062215818388157
      )
    ),
    row.names = c(NA, -5L),
    class = c("tbl_df", "tbl", "data.frame")
  ))

#> # A tibble: 5 x 3
#>   study estimate std.error
#>   <chr>    <dbl>     <dbl>
#> 1 1        0.382    0.0466
#> 2 2        0.781    0.0330
#> 3 3        0.426    0.0363
#> 4 4        0.558    0.0481
#> 5 5        0.956    0.0622

# to get dataframe
bf_meta_random(
  data = df,
  k = 3,
  iter = 1500,
  # customizing analysis with additional arguments
  control = list(max_treedepth = 15),
  output = "dataframe"
)

#> Warning: There were 2 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> # A tibble: 2 x 11
#>   term   mean    sd `2.5%` `50%` `97.5%` hpd95_lower hpd95_upper n_eff  Rhat
#>   <chr> <dbl> <dbl>  <dbl> <dbl>   <dbl>       <dbl>       <dbl> <dbl> <dbl>
#> 1 d     0.535 0.119  0.273 0.544   0.751       0.287       0.762  682.  1.00
#> 2 tau   0.277 0.118  0.133 0.249   0.602       0.119       0.518  431.  1.00
#>    bf10
#>   <dbl>
#> 1  28.2
#> 2  28.2

# }