Bayesian one-way analysis of variance

Bayesian one-way analysis of variance

bf_oneway_anova(
  data,
  x,
  y,
  bf.prior = 0.707,
  caption = NULL,
  output = "results",
  paired = FALSE,
  k = 2L,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

caption
Text to display as caption (will be displayed on top of the Bayes Factor caption/message).

output
Can either be "null" (or "caption" or "H0" or "h0"), which will return expression with evidence in favor of the null hypothesis, or "alternative" (or "title" or "H1" or "h1"), which will return expression with evidence in favor of the alternative hypothesis, or "results", which will return a dataframe with results all the details).

paired
Decides whether the design is repeated measures or not (Default: FALSE).

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Additional arguments passed to parameters::model_parameters.BFBayesFactor().

See also

Examples

# setup
set.seed(123)

# between-subjects -------------------------------------------------------
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8
)

#> # A tibble: 6 x 17
#>   term               estimate conf.low conf.high    pd rope.percentage
#>   <chr>                 <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 mu                   5.84    5.77        5.91  1               0    
#> 2 Species-setosa      -0.828  -0.927      -0.736 1               0    
#> 3 Species-versicolor   0.0920 -0.00575     0.185 0.937           0.532
#> 4 Species-virginica    0.737   0.638       0.827 1               0    
#> 5 sig2                 0.268   0.213       0.314 1               0    
#> 6 g_Species            2.22    0.290       8.64  1               0    
#>   prior.distribution prior.location prior.scale effects component      bf10
#>   <chr>                       <dbl>       <dbl> <chr>   <chr>         <dbl>
#> 1 NA                             NA          NA fixed   extra       2.03e28
#> 2 NA                             NA          NA fixed   conditional 2.03e28
#> 3 NA                             NA          NA fixed   conditional 2.03e28
#> 4 NA                             NA          NA fixed   conditional 2.03e28
#> 5 NA                             NA          NA fixed   extra       2.03e28
#> 6 NA                             NA          NA fixed   extra       2.03e28
#>       bf01 log_e_bf10 log_e_bf01 log_10_bf10 log_10_bf01
#>      <dbl>      <dbl>      <dbl>       <dbl>       <dbl>
#> 1 4.93e-29       65.2      -65.2        28.3       -28.3
#> 2 4.93e-29       65.2      -65.2        28.3       -28.3
#> 3 4.93e-29       65.2      -65.2        28.3       -28.3
#> 4 4.93e-29       65.2      -65.2        28.3       -28.3
#> 5 4.93e-29       65.2      -65.2        28.3       -28.3
#> 6 4.93e-29       65.2      -65.2        28.3       -28.3

# within-subjects -------------------------------------------------------
bf_oneway_anova(
  data = bugs_long,
  x = condition,
  y = desire,
  paired = TRUE
)

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> # A tibble: 96 x 17
#>    term           estimate conf.low conf.high    pd rope.percentage
#>    <chr>             <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#>  1 mu                6.88    6.51      7.25   1              0     
#>  2 condition-HDHF    0.905   0.612     1.19   1              0     
#>  3 condition-HDLF   -0.192  -0.466     0.0926 0.87           0.271 
#>  4 condition-LDHF    0.458   0.168     0.733  0.996          0     
#>  5 condition-LDLF   -1.17   -1.46     -0.897  1              0     
#>  6 rowid-1           0.677  -0.680     2.09   0.782          0.0741
#>  7 rowid-3           1.50    0.0916    2.92   0.948          0     
#>  8 rowid-4           0.490  -0.987     1.86   0.716          0.0890
#>  9 rowid-5          -0.822  -2.20      0.645  0.814          0.0747
#> 10 rowid-6          -2.85   -4.36     -1.47   1.00           0     
#>    prior.distribution prior.location prior.scale effects component          bf10
#>    <chr>                       <dbl>       <dbl> <chr>   <chr>             <dbl>
#>  1 NA                             NA          NA fixed   extra       1386484963.
#>  2 NA                             NA          NA fixed   conditional 1386484963.
#>  3 NA                             NA          NA fixed   conditional 1386484963.
#>  4 NA                             NA          NA fixed   conditional 1386484963.
#>  5 NA                             NA          NA fixed   conditional 1386484963.
#>  6 NA                             NA          NA random  conditional 1386484963.
#>  7 NA                             NA          NA random  conditional 1386484963.
#>  8 NA                             NA          NA random  conditional 1386484963.
#>  9 NA                             NA          NA random  conditional 1386484963.
#> 10 NA                             NA          NA random  conditional 1386484963.
#>        bf01 log_e_bf10 log_e_bf01 log_10_bf10 log_10_bf01
#>       <dbl>      <dbl>      <dbl>       <dbl>       <dbl>
#>  1 7.21e-10       21.1      -21.1        9.14       -9.14
#>  2 7.21e-10       21.1      -21.1        9.14       -9.14
#>  3 7.21e-10       21.1      -21.1        9.14       -9.14
#>  4 7.21e-10       21.1      -21.1        9.14       -9.14
#>  5 7.21e-10       21.1      -21.1        9.14       -9.14
#>  6 7.21e-10       21.1      -21.1        9.14       -9.14
#>  7 7.21e-10       21.1      -21.1        9.14       -9.14
#>  8 7.21e-10       21.1      -21.1        9.14       -9.14
#>  9 7.21e-10       21.1      -21.1        9.14       -9.14
#> 10 7.21e-10       21.1      -21.1        9.14       -9.14
#> # ... with 86 more rows

# expression -------------------------------------------------------
bf_oneway_anova(
  data = bugs_long,
  x = condition,
  y = desire,
  paired = TRUE,
  output = "null"
)

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> Multiple `BFBayesFactor` models detected - posteriors are extracted from the first numerator model.
#> See help("get_parameters", package = "insight").

#> atop(displaystyle(NULL), expr = paste("log"["e"], "(BF"["01"], 
#>     ") = ", "-21.04"))