Bayes Factor for t-test

Bayes Factor for t-test

bf_ttest(
  data,
  x,
  y = NULL,
  test.value = 0,
  paired = FALSE,
  bf.prior = 0.707,
  caption = NULL,
  output = "results",
  k = 2L,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
Either the grouping variable from the dataframe data if it's a two-sample t-test or a numeric variable if it's a one-sample t-test.

y
The column in data containing the response (outcome) variable to be plotted on the y-axis. Can be entered either as a character string (e.g., "y") or as a bare expression (e.g, y).

test.value
A number specifying the value of the null hypothesis (Default: 0).

paired
Decides whether the design is repeated measures or not (Default: FALSE).

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

caption
Text to display as caption (will be displayed on top of the Bayes Factor caption/message).

output
Can either be "null" (or "caption" or "H0" or "h0"), which will return expression with evidence in favor of the null hypothesis, or "alternative" (or "title" or "H1" or "h1"), which will return expression with evidence in favor of the alternative hypothesis, or "results", which will return a dataframe with results all the details).

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Additional arguments passed to parameters::model_parameters.BFBayesFactor().

Details

If y is NULL, a one-sample t-test will be carried out, otherwise a two-sample t-test will be carried out.

See also

Examples

# ------------------- two-samples tests -----------------------------------

# for reproducibility
set.seed(123)
library(tidyBF)

# to get dataframe
bf_ttest(
  data = mtcars,
  x = am,
  y = wt,
  paired = FALSE,
  bf.prior = 0.880
)

#> # A tibble: 1 x 17
#>   term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference     1.27    0.846      1.71     1               0
#>   prior.distribution prior.location prior.scale effects component    bf10
#>   <chr>                       <dbl>       <dbl> <chr>   <chr>       <dbl>
#> 1 cauchy                          0        0.88 fixed   conditional 1579.
#>       bf01 log_e_bf10 log_e_bf01 log_10_bf10 log_10_bf01
#>      <dbl>      <dbl>      <dbl>       <dbl>       <dbl>
#> 1 0.000633       7.36      -7.36        3.20       -3.20

# ------------------- one-samples test -----------------------------------

# to get dataframe
bf_ttest(
  data = iris,
  x = Sepal.Length,
  test.value = 5.85
)

#> # A tibble: 1 x 17
#>   term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference    0.692    0.572     0.834     1               0
#>   prior.distribution prior.location prior.scale effects component     bf10  bf01
#>   <chr>                       <dbl>       <dbl> <chr>   <chr>        <dbl> <dbl>
#> 1 cauchy                          0       0.707 fixed   conditional 0.0914  10.9
#>   log_e_bf10 log_e_bf01 log_10_bf10 log_10_bf01
#>        <dbl>      <dbl>       <dbl>       <dbl>
#> 1      -2.39       2.39       -1.04        1.04