Bayesian one-way analysis of variance

Bayesian one-way analysis of variance

bf_oneway_anova(
  data,
  x,
  y,
  bf.prior = 0.707,
  caption = NULL,
  output = "dataframe",
  paired = FALSE,
  k = 2L,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

caption
Text to display as caption (will be displayed on top of the Bayes Factor caption/message).

output
Can either be "null" (or "caption" or "H0" or "h0"), which will return expression with evidence in favor of the null hypothesis, or "alternative" (or "title" or "H1" or "h1"), which will return expression with evidence in favor of the alternative hypothesis, or "dataframe", which will return a dataframe with results all the details).

paired
Decides whether the design is repeated measures or not (Default: FALSE).

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Additional arguments passed to parameters::model_parameters.BFBayesFactor().

See also

Examples

# setup
set.seed(123)

# between-subjects -------------------------------------------------------
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8
)

#> # A tibble: 6 x 13
#>   term               estimate conf.low conf.high    pd rope.percentage
#>   <chr>                 <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 mu                   5.84    5.77        5.91  1               0    
#> 2 Species-setosa      -0.829  -0.927      -0.736 1               0    
#> 3 Species-versicolor   0.0917 -0.00575     0.185 0.937           0.434
#> 4 Species-virginica    0.738   0.638       0.827 1               0    
#> 5 sig2                 0.267   0.213       0.314 1               0    
#> 6 g_Species            2.21    0.290       8.64  1               0    
#>   prior.distribution prior.location prior.scale effect component      bf10
#>   <chr>                       <dbl>       <dbl> <chr>  <chr>         <dbl>
#> 1 NA                             NA          NA fixed  extra       2.03e28
#> 2 NA                             NA          NA fixed  conditional 2.03e28
#> 3 NA                             NA          NA fixed  conditional 2.03e28
#> 4 NA                             NA          NA fixed  conditional 2.03e28
#> 5 NA                             NA          NA fixed  extra       2.03e28
#> 6 NA                             NA          NA fixed  extra       2.03e28
#>   log_e_bf10
#>        <dbl>
#> 1       65.2
#> 2       65.2
#> 3       65.2
#> 4       65.2
#> 5       65.2
#> 6       65.2

# within-subjects -------------------------------------------------------
bf_oneway_anova(
  data = bugs_long,
  x = condition,
  y = desire,
  paired = TRUE
)

#> # A tibble: 96 x 13
#>    term           estimate conf.low conf.high    pd rope.percentage
#>    <chr>             <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#>  1 mu                6.88    6.50      7.24   1              0     
#>  2 condition-HDHF    0.912   0.612     1.19   1              0     
#>  3 condition-HDLF   -0.191  -0.467     0.0988 0.864          0.727 
#>  4 condition-LDHF    0.457   0.172     0.735  0.994          0.124 
#>  5 condition-LDLF   -1.17   -1.46     -0.889  1              0     
#>  6 rowid-1           0.681  -0.717     2.09   0.780          0.209 
#>  7 rowid-3           1.48    0.0826    2.92   0.954          0.0396
#>  8 rowid-4           0.486  -1.00      1.86   0.712          0.251 
#>  9 rowid-5          -0.811  -2.17      0.661  0.82           0.177 
#> 10 rowid-6          -2.87   -4.36     -1.45   1.00           0     
#>    prior.distribution prior.location prior.scale effect component          bf10
#>    <chr>                       <dbl>       <dbl> <chr>  <chr>             <dbl>
#>  1 NA                             NA          NA fixed  extra       1381847039.
#>  2 NA                             NA          NA fixed  conditional 1381847039.
#>  3 NA                             NA          NA fixed  conditional 1381847039.
#>  4 NA                             NA          NA fixed  conditional 1381847039.
#>  5 NA                             NA          NA fixed  conditional 1381847039.
#>  6 NA                             NA          NA random conditional 1381847039.
#>  7 NA                             NA          NA random conditional 1381847039.
#>  8 NA                             NA          NA random conditional 1381847039.
#>  9 NA                             NA          NA random conditional 1381847039.
#> 10 NA                             NA          NA random conditional 1381847039.
#>    log_e_bf10
#>         <dbl>
#>  1       21.0
#>  2       21.0
#>  3       21.0
#>  4       21.0
#>  5       21.0
#>  6       21.0
#>  7       21.0
#>  8       21.0
#>  9       21.0
#> 10       21.0
#> # ... with 86 more rows

# expression -------------------------------------------------------
bf_oneway_anova(
  data = bugs_long,
  x = condition,
  y = desire,
  paired = TRUE,
  output = "null"
)

#> atop(displaystyle(NULL), expr = paste("log"["e"], "(BF"["01"], 
#>     ") = ", "-21.04"))