Bayesian one-way analysis of variance

Bayesian one-way analysis of variance

bf_oneway_anova(
  data,
  x,
  y,
  bf.prior = 0.707,
  caption = NULL,
  output = "results",
  paired = FALSE,
  k = 2,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

caption
Text to display as caption (will be displayed on top of the Bayes Factor caption/message).

output
Can either be "null" (or "caption" or "H0" or "h0"), which will return expression with evidence in favor of the null hypothesis, or "alternative" (or "title" or "H1" or "h1"), which will return expression with evidence in favor of the alternative hypothesis, or "results", which will return a dataframe with results all the details).

paired
if TRUE, observations are paired

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

...
Additional arguments.

See also

Examples

# \donttest{
# setup
set.seed(123)

# between-subjects -------------------------------------------------------
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8
)

#> # A tibble: 1 x 8
#>      bf10        error     bf01 log_e_bf10 log_e_bf01 log_10_bf10 log_10_bf01
#>     <dbl>        <dbl>    <dbl>      <dbl>      <dbl>       <dbl>       <dbl>
#> 1 2.03e28 0.0000000478 4.93e-29       65.2      -65.2        28.3       -28.3
#>   bf.prior
#>      <dbl>
#> 1      0.8

# within-subjects -------------------------------------------------------
bf_oneway_anova(
  data = bugs_long,
  x = condition,
  y = desire,
  paired = TRUE
)

#> # A tibble: 1 x 8
#>          bf10   error     bf01 log_e_bf10 log_e_bf01 log_10_bf10 log_10_bf01
#>         <dbl>   <dbl>    <dbl>      <dbl>      <dbl>       <dbl>       <dbl>
#> 1 1372773375. 0.00701 7.28e-10       21.0      -21.0        9.14       -9.14
#>   bf.prior
#>      <dbl>
#> 1    0.707

# }