Bayes Factor for one-way analysis of variance (ANOVA)

Bayes Factor for one-way analysis of variance (ANOVA)

bf_oneway_anova(
  data,
  x,
  y,
  subject.id = NULL,
  paired = FALSE,
  bf.prior = 0.707,
  top.text = NULL,
  output = "dataframe",
  k = 2L,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

subject.id
In case of repeated measures design (paired = TRUE, i.e.), this argument specifies the subject or repeated measures id. Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate.

paired
Decides whether the design is repeated measures or not (Default: FALSE).

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

top.text
Text to display as top.text (will be displayed on top of the Bayes Factor top.text/message).

output
If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Additional arguments passed to parameters::model_parameters.BFBayesFactor().

Note

If you want to set paired = TRUE, you are going to need to install the development version of BayesFactor (0.9.12-4.3). You can download it by running: remotes::install_github("richarddmorey/BayesFactor/pkg/BayesFactor").

See also

Examples

# setup
set.seed(123)

# between-subjects -------------------------------------------------------

# to get dataframe
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8,
  output = "dataframe"
)

#> # A tibble: 6 x 18
#>   term               estimate conf.low conf.high    pd rope.percentage
#>   <chr>                 <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 mu                   5.84     5.76       5.93  1               0    
#> 2 Species-setosa      -0.829   -0.943     -0.708 1               0    
#> 3 Species-versicolor   0.0917  -0.0277     0.208 0.937           0.434
#> 4 Species-virginica    0.738    0.619      0.852 1               0    
#> 5 sig2                 0.267    0.210      0.335 1               0    
#> 6 g_Species            2.21     0.290     15.5   1               0    
#>   prior.distribution prior.location prior.scale effect component      bf10
#>   <chr>                       <dbl>       <dbl> <chr>  <chr>         <dbl>
#> 1 NA                             NA          NA fixed  extra       2.03e28
#> 2 NA                             NA          NA fixed  conditional 2.03e28
#> 3 NA                             NA          NA fixed  conditional 2.03e28
#> 4 NA                             NA          NA fixed  conditional 2.03e28
#> 5 NA                             NA          NA fixed  extra       2.03e28
#> 6 NA                             NA          NA fixed  extra       2.03e28
#>   log_e_bf10    r2 std.dev ci.width r2.conf.low r2.conf.high
#>        <dbl> <dbl>   <dbl>    <dbl>       <dbl>        <dbl>
#> 1       65.2 0.612  0.0310       95       0.545        0.668
#> 2       65.2 0.612  0.0310       95       0.545        0.668
#> 3       65.2 0.612  0.0310       95       0.545        0.668
#> 4       65.2 0.612  0.0310       95       0.545        0.668
#> 5       65.2 0.612  0.0310       95       0.545        0.668
#> 6       65.2 0.612  0.0310       95       0.545        0.668

# within-subjects -------------------------------------------------------

# to get expression (needs `BayesFactor 0.9.12-4.3`)
if (utils::packageVersion("BayesFactor") >= package_version("0.9.12-4.3")) {
  bf_oneway_anova(
    data = bugs_long,
    x = condition,
    y = desire,
    subject.id = subject,
    paired = TRUE,
    output = "expression"
  )
}

#> paste("log"["e"], "(BF"["01"], ") = ", "-21.04", ", ", widehat(italic(R^"2"))["median"]^"posterior", 
#>     " = ", "0.53", ", CI"["95%"]^"HDI", " [", "0.46", ", ", "0.59", 
#>     "]", ", ", italic("r")["Cauchy"]^"JZS", " = ", "0.71")